backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic[email]==2.4.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
celery==5.3.4
redis==5.0.1
twilio==8.10.0
requests==2.31.0
google-generativeai==0.3.1
